+++ package.json
@@ -20,7 +20,7 @@
     "styled-components": "^5.3.6",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4",
-    "yup": "^1.0.2"
+    "zod": "^3.21.4"
   },
   "scripts": {
     "start": "craco start",
