+++ src/sandbox/jakub/index.tsx
@@ -1,9 +1,12 @@
 import React from 'react'
 
+import LoginPage from './login'
+
 const JakubSandboxPage: React.FC = () => {
   return (
     <>
       <h1>Jakub Sandbox</h1>
+      <LoginPage />
     </>
   )
 }
