+++ src/sandbox/jakub/login/schema.ts
@@ -1,8 +1,12 @@
-import * as y from 'yup'
+import { z } from 'zod'
 
-export const LoginFormSchema = y.object({
-  email: y.string().email().required(),
-  password: y.string().required(),
+export const LoginFormSchema = z.object({
+  email: z
+    .string({ required_error: 'This is required' })
+    .email({ message: 'Must be a valid email' }),
+  password: z
+    .string({ required_error: 'This is required' })
+    .min(8, { message: 'Too short - min. 8 characters' }),
 })
 
-export type LoginForm = y.Asserts<typeof LoginFormSchema>
+export type LoginForm = z.infer<typeof LoginFormSchema>
