+++ src/sandbox/jakub/login/utils.ts
@@ -1,0 +1,5 @@
+import { LoginForm } from './schema'
+
+export const getLoginFormDefaultValues = (): LoginForm => {
+  return { email: '', password: '' }
+}
